# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KRA, KPI & Task Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /kras:
    get:
      operationId: listKras
      tags: [kras]
      summary: List KRAs
      parameters:
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of KRAs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Kra"
    post:
      operationId: createKra
      tags: [kras]
      summary: Create a KRA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KraInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kra"
  /kras/{id}:
    get:
      operationId: getKra
      tags: [kras]
      summary: Get a KRA
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: KRA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kra"
    patch:
      operationId: updateKra
      tags: [kras]
      summary: Update a KRA
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KraUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kra"
    delete:
      operationId: deleteKra
      tags: [kras]
      summary: Delete a KRA
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /kras/{id}/score:
    patch:
      operationId: scoreKra
      tags: [kras]
      summary: Update KRA achievement score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KraScoreInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kra"
  /kras/{id}/submit:
    patch:
      operationId: submitKraForClosure
      tags: [kras]
      summary: Employee submits KRA for closure
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kra"
  /kras/{id}/hr-approve:
    patch:
      operationId: hrApproveKra
      tags: [kras]
      summary: HR department approves or rejects a new KRA
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HrApproveKraInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kra"
  /kras/{id}/approve-close:
    patch:
      operationId: approveKraClosure
      tags: [kras]
      summary: Manager or HOD approves/rejects KRA closure
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveCloseInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kra"

  # ── KPIs ──────────────────────────────────────────────────────────────────────
  /kpis:
    get:
      operationId: listKpis
      tags: [kpis]
      summary: List KPIs
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of KPIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Kpi"
    post:
      operationId: createKpi
      tags: [kpis]
      summary: Create a KPI record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KpiInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kpi"
  /kpis/{id}:
    get:
      operationId: getKpi
      tags: [kpis]
      summary: Get a KPI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: KPI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kpi"
    patch:
      operationId: updateKpi
      tags: [kpis]
      summary: Update a KPI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KpiUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kpi"
    delete:
      operationId: deleteKpi
      tags: [kpis]
      summary: Delete a KPI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Tasks ─────────────────────────────────────────────────────────────────────
...
  /kpis/calculate:
    post:
      operationId: calculateKpi
      tags: [kpis]
      summary: Auto-calculate KPI from KRA achievement and task completion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KpiCalculateInput"
      responses:
        "201":
          description: Calculated/updated KPI record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kpi"
components:
  schemas:
...
    Kra:
      type: object
      required: [id, title, weightage, departmentId, kraStatus, createdAt]
      properties:
        id:             { type: integer }
        title:          { type: string }
        description:    { type: ["string", "null"] }
        weightage:      { type: number }
        achievementPct: { type: ["number", "null"] }
        departmentId:   { type: integer }
        departmentName: { type: string }
        employeeId:     { type: ["integer", "null"] }
        employeeName:   { type: ["string", "null"] }
        reviewPeriod:     { type: string, enum: [monthly, quarterly, yearly] }
        frequency:        { type: string, enum: [daily, weekly, bi_weekly, monthly, quarterly, yearly] }
        dueDate:          { type: ["string", "null"] }
        kraStatus:        { type: string, enum: [active, submitted, manager_approved, approved, rejected] }
        hrApprovalStatus: { type: string, enum: [pending_hr, hr_approved, hr_rejected] }
        submittedAt:      { type: ["string", "null"] }
        closedAt:         { type: ["string", "null"] }
        createdAt:        { type: string }
    KraInput:
      type: object
      required: [title, weightage, departmentId, frequency]
      properties:
        title:        { type: string, minLength: 1 }
        description:  { type: string }
        weightage:    { type: number }
        departmentId: { type: integer }
        employeeId:   { type: integer }
        frequency:    { type: string, enum: [daily, weekly, bi_weekly, monthly, quarterly, yearly] }
        dueDate:      { type: string }
    KraUpdate:
      type: object
      properties:
        title:          { type: string }
        description:    { type: string }
        weightage:      { type: number }
        achievementPct: { type: number }
        reviewPeriod:   { type: string, enum: [monthly, quarterly, yearly] }
        frequency:      { type: string, enum: [daily, weekly, bi_weekly, monthly, quarterly, yearly] }
        dueDate:        { type: string }
        employeeId:     { type: integer }
    HrApproveKraInput:
      type: object
      required: [approved]
      properties:
        approved: { type: boolean }
    KraScoreInput:
      type: object
      required: [achievementPct]
      properties:
        achievementPct: { type: number }

    # ── KPI ─────────────────────────────────────────────────────────────────────
    Kpi:
      type: object
      required: [id, employeeId, month, year, kraAchievement, taskCompletion, productivity, punctuality, discipline, totalScore, rating, createdAt]
      properties:
        id:              { type: integer }
        employeeId:      { type: integer }
        employeeName:    { type: string }
        departmentId:    { type: integer }
        departmentName:  { type: string }
        month:           { type: integer }
        year:            { type: integer }
        kraAchievement:  { type: number }
        taskCompletion:  { type: number }
        productivity:    { type: number }
        punctuality:     { type: number }
        discipline:      { type: number }
        totalScore:      { type: number }
        rating:          { type: string }
        createdAt:       { type: string }
    KpiInput:
      type: object
      required: [employeeId, month, year, kraAchievement, taskCompletion, productivity, punctuality, discipline]
      properties:
        employeeId:     { type: integer }
        month:          { type: integer }
        year:           { type: integer }
        kraAchievement: { type: number }
        taskCompletion: { type: number }
        productivity:   { type: number }
        punctuality:    { type: number }
        discipline:     { type: number }
    KpiUpdate:
      type: object
      properties:
        kraAchievement: { type: number }
        taskCompletion: { type: number }
        productivity:   { type: number }
        punctuality:    { type: number }
        discipline:     { type: number }

    # ── Task ────────────────────────────────────────────────────────────────────
...
    ApproveCloseInput:
      type: object
      required: [approved]
      properties:
        approved: { type: boolean }
        reason:   { type: string }
...
    KpiCalculateInput:
      type: object
      required: [employeeId, month, year]
      properties:
        employeeId: { type: integer }
        month:      { type: integer }
        year:       { type: integer }